moving logger code
